particlestuff, fixes

+++ src/main/java/techguns/entities/projectiles/AdvancedBulletProjectile.java
@@ -10,7 +10,7 @@ public class AdvancedBulletProjectile extends GenericProjectile{
 
 
 	public AdvancedBulletProjectile(World world, EntityLivingBase p, float damage, float speed, int TTL, float spread,
-			int dmgDropStart, int dmgDropEnd, float dmgMin, float penetration, boolean blockdamage, EnumBulletFirePos leftGun) {
+			float dmgDropStart, float dmgDropEnd, float dmgMin, float penetration, boolean blockdamage, EnumBulletFirePos leftGun) {
 		super(world, p, damage, speed, TTL, spread, dmgDropStart, dmgDropEnd, dmgMin, penetration, blockdamage, leftGun);
 	}
 	
@@ -22,7 +22,7 @@ public class AdvancedBulletProjectile extends GenericProjectile{
 	public static class Factory implements IProjectileFactory<AdvancedBulletProjectile> {
 
 		@Override
-		public AdvancedBulletProjectile createProjectile(GenericGun gun, World world, EntityLivingBase p, float damage, float speed, int TTL, float spread, int dmgDropStart, int dmgDropEnd,
+		public AdvancedBulletProjectile createProjectile(GenericGun gun, World world, EntityLivingBase p, float damage, float speed, int TTL, float spread, float dmgDropStart, float dmgDropEnd,
 				float dmgMin, float penetration, boolean blockdamage, EnumBulletFirePos firePos, float radius, double gravity) {
 			return new AdvancedBulletProjectile(world,p,damage,speed,TTL,spread,dmgDropStart,dmgDropEnd,dmgMin,penetration,blockdamage,firePos);
 		}
