+++ src/main/java/techguns/entities/projectiles/TeslaProjectile.java
@@ -31,21 +31,21 @@ public class TeslaProjectile extends AbstractBeamProjectile{
 	}
 	
 //	public TeslaProjectile(World worldIn, double posX, double posY, double posZ, float yaw, float pitch, float damage,
-//			float speed, int TTL, float spread, int dmgDropStart, int dmgDropEnd, float dmgMin, float penetration,
+//			float speed, int TTL, float spread, float dmgDropStart, float dmgDropEnd, float dmgMin, float penetration,
 //			boolean blockdamage, boolean leftGun) {
 //		super(worldIn, posX, posY, posZ, yaw, pitch, damage, speed, TTL, spread, dmgDropStart, dmgDropEnd, dmgMin, penetration,
 //				blockdamage, leftGun);
 //	}
 	
 	public TeslaProjectile(World par2World, EntityLivingBase p, float damage, float speed, int TTL, float spread,
-			int dmgDropStart, int dmgDropEnd, float dmgMin, float penetration, boolean blockdamage, EnumBulletFirePos leftGun, int chainTargets) {
+			float dmgDropStart, float dmgDropEnd, float dmgMin, float penetration, boolean blockdamage, EnumBulletFirePos leftGun, int chainTargets) {
 		super(par2World, p, damage, speed, TTL, spread, dmgDropStart, dmgDropEnd, dmgMin, penetration, blockdamage, leftGun);
 		this.chainTargets = chainTargets;
 		trace();
 	}
 	
-	public TeslaProjectile(World world, EntityLivingBase shooter, Entity source, EntityLivingBase target, int chainTargets, float damage, float speed, int TTL, int dmgDropStart,
-			int dmgDropEnd, float dmgMin, float penetration, boolean blockdamage) {
+	public TeslaProjectile(World world, EntityLivingBase shooter, Entity source, EntityLivingBase target, int chainTargets, float damage, float speed, int TTL, float dmgDropStart,
+			float dmgDropEnd, float dmgMin, float penetration, boolean blockdamage) {
 		super(world, shooter, damage, speed, TTL, 0, dmgDropStart, dmgDropEnd, dmgMin, penetration, blockdamage, EnumBulletFirePos.CENTER);
 		maxTicks = (short) this.ticksToLive;
 		this.chainTargets = chainTargets;
@@ -142,7 +142,7 @@ public class TeslaProjectile extends AbstractBeamProjectile{
 	public static class Factory implements IProjectileFactory<TeslaProjectile> {
 		@Override
 		public TeslaProjectile createProjectile(GenericGun gun, World world, EntityLivingBase p, float damage, float speed,
-				int TTL, float spread, int dmgDropStart, int dmgDropEnd, float dmgMin, float penetration,
+				int TTL, float spread, float dmgDropStart, float dmgDropEnd, float dmgMin, float penetration,
 				boolean blockdamage, EnumBulletFirePos firePos, float radius, double gravity) {
 			return new TeslaProjectile(world, p, damage, speed, TeslaProjectile.TTL, spread, dmgDropStart, dmgDropEnd, dmgMin, penetration, blockdamage, firePos, CHAIN_TARGETS);
 		}
